Level Editor Tool (#98)
* Created Dedicated Folders and Script for Level Editor

Added a "Tools" Folder and an "Editor" folder to not mess with building later down the line. Started programming the tool as well.

* Update On LevelEditor Script

Current Progress. Need to fix it so that it detects input in the scene view.

* Update to LevelEditor Script

Added functionality so that it spawns to where you click in the scene view. Next I need to work on it being positioned so that it is directly on top. I also need to make it snap to the grid.

* Added Undo Functionality to Level Editor

* Updated LevelEditor Script & Creation of Optimization

* Updated Merge Meshes Script

* Finished MergeMeshesManager script

* Move Folders and Add Eraser Feature

Moved folders around and created an eraser feature for the level editor. When the eraser tool is being used, you can click prefabs in the scene to delete it.

* Added PrefabPreview and Escape Keybind

* Removed Debug Statements