Stab and dash, file moving, and player movement adjustments (#104)
* Test Scene Added

* Added Stab

Stab Added as a script in a temporary player 3d object.

Note: Stab is separate from dash. I'm trying to keep them as two scripts so they function without the other, but still need to test this.

* Dash added, Sword script added

First version of Dash.
Sword script checks collision to a child object of player. This will cause the character controller to move forward.

Still missing interruption and improvement of movement and detection. A very hacked in presentation is used right now.

* Minor Changes

* sword following behavior

* adjusted test scene to detect collision

* scripts clean up

* file, prefab, script organization

* jump, demon sword following adjustments

* basic dash implementation

* adjusted drag implementation and stab dashing

* basic functional stab and dash implementation

* stab and dash implementation complete

known bug
- dashing imediatly after stab -> dash prevents player inputs

* added gravity accel fixed stab and dash bugs

- fixed dash 

- fixed bug of immediately after stab -> dash

---------

Co-authored-by: Pedro P <pedroaperezdr@gmail.com>